# inbounds for edge iteration and graph construction