```
g = CompleteGraph(10_000)

function bar(x)
    a = 0
    for v in x
        a += 1
    end
    return a
end
```

current master:
```
julia> @benchmark bar(edges(g))
BenchmarkTools.Trial:
  memory estimate:  32 bytes
  allocs estimate:  2
  --------------
  minimum time:     612.813 ms (0.00% GC)
  median time:      624.377 ms (0.00% GC)
  mean time:        629.321 ms (0.00% GC)
  maximum time:     658.196 ms (0.00% GC)
  --------------
  samples:          8
  evals/sample:     1
```

This PR:
```
julia> @benchmark bar(edges(g))
BenchmarkTools.Trial:
  memory estimate:  32 bytes
  allocs estimate:  2
  --------------
  minimum time:     547.716 ms (0.00% GC)
  median time:      559.328 ms (0.00% GC)
  mean time:        576.167 ms (0.00% GC)
  maximum time:     648.430 ms (0.00% GC)
  --------------
  samples:          9
  evals/sample:     1
```