@jpfairbanks 2017-12-05
This requires that any type `<:AbstractSimpleGraph` support guarantee validity of all indices between `1:nv(g)`. I guess that makes sense, but we would be enforcing that as part of this change.