@sbromberger 2017-12-06
> This requires that any type <:AbstractSimpleGraph support guarantee validity of all indices between 1:nv(g)

Great point, but this is already a promise, since vertices are a `UnitRange`. Do you see any other issues with this PR?